Deprecate url methods and call in new location
