+++ solr/solrj/src/test/org/apache/solr/common/util/URLUtilTest.java
@@ -16,6 +16,11 @@
  */
 package org.apache.solr.common.util;
 
+import static org.apache.solr.common.util.URLUtil.getNodeNameFromSolrUrl;
+import static org.apache.solr.common.util.Utils.getBaseUrlForNodeName;
+
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
 import org.apache.solr.SolrTestCase;
 import org.junit.Test;
 
@@ -94,5 +99,30 @@ public class URLUtilTest extends SolrTestCase {
     assertEquals(
         "http://localhost:8983/solr/sTrAnGe-name.for_core",
         URLUtil.buildCoreUrl("http://localhost:8983/solr", "sTrAnGe-name.for_core"));
+  }
+
+  @Test
+  public void testGetNodeNameFromSolrUrl() throws MalformedURLException, URISyntaxException {
+    assertEquals("node-1-url:8983_solr", getNodeNameFromSolrUrl("https://node-1-url:8983/solr"));
+    assertEquals("node-1-url:8983_solr", getNodeNameFromSolrUrl("http://node-1-url:8983/solr"));
+    assertEquals("node-1-url:8983_api", getNodeNameFromSolrUrl("http://node-1-url:8983/api"));
+    assertThrows(MalformedURLException.class, () -> getNodeNameFromSolrUrl("node-1-url:8983/solr"));
+    assertThrows(
+        URISyntaxException.class, () -> getNodeNameFromSolrUrl("http://node-1-url:8983/solr^"));
+  }
+
+  @Test
+  public void testGetBaseUrlForNodeName() {
+    assertEquals(
+        "http://app-node-1:8983/solr",
+        URLUtil.getBaseUrlForNodeName("app-node-1:8983_solr", "http", false));
+    assertEquals(
+        "https://app-node-1:8983/solr",
+        getBaseUrlForNodeName("app-node-1:8983_solr", "https", false));
+    assertEquals(
+        "http://app-node-1:8983/api", getBaseUrlForNodeName("app-node-1:8983_solr", "http", true));
+    assertEquals(
+        "https://app-node-1:8983/api",
+        getBaseUrlForNodeName("app-node-1:8983_solr", "https", true));
   }
 }
